M44-9 move sidebar related logic into its own component
